+++ src/webviews/lwc-ui-panel.ts
@@ -114,8 +114,8 @@ export class LwcUiPanel {
   }
 
   public asWebviewUri(path: string[]): string {
-    path.unshift("out","resources","webviews");
-    const resourceUri = vscode.Uri.joinPath(this.extensionUri, ...path);
+    const pathFull = ["out", "resources", "webviews", ...path]
+    const resourceUri = vscode.Uri.joinPath(this.extensionUri, ...pathFull);
     return this.panel.webview.asWebviewUri(resourceUri).toString();
   }