+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Display Git & ticketing providers icons
   - Add button to view pipeline configuration from pipeline view
   - Allow more formats of jiraHost
+  - Fix issue when both Jira PAT and Email + Token are set
 
 ## [6.14.5] 2024-10-22